requirements.txt
# Core AI and Agent Framework
pyautogen

# Git integration
GitPython

# Web scraping and HTTP requests
requests
beautifulsoup4

# Environment variable management
python-dotenv

# Search API integration (will be used for web search)
google-search-results

# PDF generation and manipulation (for future phases)
reportlab
PyPDF2

# Data handling and utilities
pydantic

# Phase 8 - Advanced Database Integration
asyncpg>=0.28.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0
redis>=4.5.0
chromadb>=0.4.0
aiosqlite>=0.19.0

# Phase 9 - Payment & Subscription System
stripe>=6.0.0

# Phase 10 - Advanced AI Features & Analytics
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0

# Additional FastAPI dependencies for enterprise features
uvicorn[standard]>=0.23.0
fastapi>=0.100.0
websockets>=11.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0